change Makefiles and remove lnsrc/ulnsrc
Since now -I. is included in CFLAGS in the Makefiles, we got a
problem with the use of <memory.h>.

We could rename memory.[ch] to simmem.[ch].

But for now, I either deleted the inclusion of <memory.h> or
changed it to <string.h> in all files, since as of C89 all the
mem* functions are in <string.h>.

That's probably the reason while "-iquote ." is used in the GNUmakefiles.

disas.c is now included in mosteksim. This is temporary, until
the move of the ICE into z80core.