z80asm: use BYTE & WORD, don't use sprintf, do own %x, REG* again
This is a rather large commit, since my playing around with
Coherent 3.2 unearthed some problems with the code.

1. Expression evaluation needs to be done with 16-bit unsigned
   arithmetic, otherwise one gets problems like "(1 EQ 1) SHR 1"
   resulting in 0xffff and not 0x7fff.
2. Just bitten the bullet and introduced BYTE and WORD like in
   z80core and used them where applicable.
3. Listing line counting needs to be done with long's, since,
   for example, the full macro expansion listing of ex.mac has
   157226 source lines.
4. Don't use sprintf, since Coherent's is not usable. It doesn't
   return the number of characters printed.
5. Do '%x' formatting by hand, since Coherent's is upper case.
6. Couldn't stop and have redone the register/flags number
   assignments so that "<< 3" shifting is no longer needed.

Now the fully macro expanded listing of ex.mac is identical on
Coherent 3.2 and Linux. The stripped binary of z80asm on
Coherent is only 2.5K larger than the old version in /usr/local/bin.
Probably due to my relentless fiddling with z80arfun.c.