z80asm: limits part 3, move strsave to z80amain.c
Remove arbitrary file name length limit and source file count limit.

Move strsave to z80amain.c and add fatal call if out of memory.

There is a change in semantics in that object and list files
specified on the command line with an extension now keep there
extension.

This situation was not mentioned in README-asm.txt, so it was
undefined behaviour.

But this allows, for example, to now specify "-fb -ox.com" to
directly create a .com file.