z80core: add I/O block instructions S,H,P,N,C flags back for ex.com
I think users are not expecting that ex.com as supplied on z80tests.dsk
fails. This makes it pass.

This is slightly different than ae6346b6901dc53c46689ff0744a2ae8d1bf85d4.
The start instead of the final value of B is used for the parity flag
calculation.