make web-based frontend a runtime opt-in feature
When the web-based frontend code is compiled in, it can now be enabled
at runtime with the new option -n.

Default port number is set in sim.h, but it can be changed with a
new configuration file option "ns_port".

Added SYSDOCROOT define to Makefile. This directory will be used
when ../../webfrontend/www/<machine> can't be found.