intelmds: makes no sense
I was thinking about adding support for the 206 Controller to the
CP/M BIOS, but this is an emulator of a real machine and this
means one would also need support for bad sector mapping for the
hard drive like ISIS's fixmap.

Just forget it. If there is need to use large disks with CP/M just
use cpmsim, that's what it's there for.