intelmds: make ISIS hdcopy work
hdcopy works by formatting 4 tracks, copy them, and repeat.

This didn't work because the format track emulation would create
a new disk image file when formatting track 0 and the write emulation
checks that the disk image file has the full size.

Add an ftruncate to make the file full size when formatting track 0.

Also expand description of what is emulated in the README.