intelmds: reduce BIOS size, proper cpm.bin, add original CP/M disk
Reduce code duplication in the BIOS. It now fits in 7 pages at starts
at F100.

Adjust the boot loader and add proper defines for the BIOS.

The file cpm.bin is now generated with movcpm.

Copy new versions of boot.asm and bios.asm to cpmgen.dsk and add
a sysgen.sub file.

Add a copy of the original Intel CP/M disk.