make some noise; remove all .o and .d files on "make clean"
Add sound support to the frontpanel library.

Add a power flag to frontpanel, so it can determine that the power is on.
Can be bound with new API call fp_bindPowerFlag().

Add option "fansound" and switch options "onsound" and "offsound" to the
frontpanel configuration file.

Change "make clean" in all Makefile's to remove all .d and .o files, so
there won't be any leftovers from previous compiles with different options.

Initialize all SDL libraries in simsdl.c in case we want to use these
not only in the frontpanel library.