clean-up netsrv and fix zero bug with message queue ids
Move system includes not needed in netsrv.h to netsrv.c. This revealed that
cromemco-dazzler and imsai-vio need to include string.h with HAS_NETSERVER.

At least on Linux the first message queue created has an id of 0,
this triped up the code which expected a value > 0. Use -1 to mean no queue
connected and not 0.

The manual page of msgget(2) also states that the returned id is
a nonnegative integer, so 0 is a completely legal value.