Update Makefiles for MacOS and Homebrew
The current Makefiles are expecting that the SDL2 libraries was
installed as a framework. This is not the case when using Homebrew
on MacOS. This commit updates the Makefiles to use the correct
library paths for MacOS and Homebrew.

Currently I have sdl2 sdl2_gfx sdl2_image sdl2_mixer sdl2_sound and
sdl2_ttf installed on my machine,but all of them might not be required.